requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Financial Data APIs
yfinance>=0.2.0
alpha-vantage>=2.3.1

# Database
psycopg2-binary>=2.9.0  # Must be installed before timescale-vector
sqlalchemy>=2.0.0
# timescale-vector>=0.0.1  # Temporarily commented - causes psycopg2 build conflict

# Data Storage
pyarrow>=12.0.0
fastparquet>=2023.4.0
duckdb>=0.8.0

# Statistical Analysis
statsmodels>=0.14.0
arch>=5.5.0  # For GARCH models

# Technical Indicators
# ta-lib>=0.4.0  # Commented out - requires C compilation, using pandas-ta instead
pandas-ta
numba
# Web Framework
flask>=2.3.0
flask-cors>=4.0.0
streamlit>=1.24.0
dash>=2.11.0
plotly>=5.15.0

# API & Security
python-dotenv>=1.0.0
pydantic>=2.0.0
cryptography>=41.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Logging & Monitoring
loguru>=0.7.0
wandb>=0.15.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
joblib>=1.3.0
rich>=13.0.0  # For terminal UI

# Reproducibility
pytorch-lightning>=2.0.0
hydra-core>=1.3.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
isort>=5.12.0

# Jupyter (for experimentation)
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
